Adjust GCS filter false positive rate based on number of elements.
The current GCS filter code is really designed exclusively for 1MB blocks. As
the number of elements in the filter goes up the false positive rate of the filter
goes up. The end result would be that filters for larger blocks would have a very
high false positive rate and trigger neutrino wallets to download nearly all blocks.

This commit adjusts the GCS filter m value upwards to ensure a consistent false
positive rate as the number of elements in the filter grows. I've selected 8,000
as the target base value as that rougly corresponds to a 1MB block containing 2,000
transactions with an average of two inputs and two outputs per tranaction. This
number may be an over estimate for the number of elements in a 1 MB block as a
block often contains multipe outputs paying the same script or spending from the
same script. In such cases the script is only added to the filter once.

Thus this commit ensures the m value will never be less than 784931 but can go
higher as the number of non-duplicate elements in the block increases.