Fix leading zero bug in hdkeychain
The private child key derivation function incorrectly does not add leading zeros
if the resultant child key is less than 32 bytes. This is not to the bip32 specifications.
This commit fixes the bug, pads with leading zeros if necessary, and adds a test
vector.

Note that this will break any wallets that used the old incorrect derivation function.
If you use this function in production, it is up to you to decide if you want
to upgrade.